+++ bittytax/conv/mergers/__init__.py
@@ -1,0 +1,3 @@
+from . import bscscan
+from . import etherscan
+from . import hecoinfo