+++ bittytax/conv/parsers/hecoinfo.py
@@ -18,7 +18,8 @@ def parse_hecoinfo(data_row, _parser, **_kwargs):
                                                  buy_quantity=row_dict['Value_IN(HT)'],
                                                  buy_asset="HT",
                                                  wallet=WALLET,
-                                                 note=row_dict.get('PrivateNote', ''))
+                                                 note=get_note(row_dict))
+
     elif Decimal(row_dict['Value_OUT(HT)']) > 0:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_WITHDRAWAL,
                                                  data_row.timestamp,
@@ -27,7 +28,7 @@ def parse_hecoinfo(data_row, _parser, **_kwargs):
                                                  fee_quantity=row_dict['TxnFee(HT)'],
                                                  fee_asset="HT",
                                                  wallet=WALLET,
-                                                 note=row_dict.get('PrivateNote', ''))
+                                                 note=get_note(row_dict))
     else:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_SPEND,
                                                  data_row.timestamp,
@@ -36,7 +37,13 @@ def parse_hecoinfo(data_row, _parser, **_kwargs):
                                                  fee_quantity=row_dict['TxnFee(HT)'],
                                                  fee_asset="HT",
                                                  wallet=WALLET,
-                                                 note=row_dict.get('PrivateNote', ''))
+                                                 note=get_note(row_dict))
+
+def get_note(row_dict):
+    if row_dict['Status'] != '':
+        return "Failure"
+
+    return row_dict.get('PrivateNote', '')
 
 def parse_hecoinfo_internal(data_row, _parser, **_kwargs):
     row_dict = data_row.row_dict
@@ -55,15 +62,16 @@ def parse_hecoinfo_internal(data_row, _parser, **_kwargs):
                                                  sell_asset="HT",
                                                  wallet=WALLET)
 
-heco1 = DataParser(DataParser.TYPE_EXPLORER,
-           "HecoInfo (HECO Transactions)",
-           ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
-            'Value_IN(HT)', 'Value_OUT(HT)', None, 'TxnFee(HT)', 'TxnFee(USD)',
-            'Historical $Price/HT', 'Status', 'ErrCode'],
-           worksheet_name="HecoInfo",
-           row_handler=parse_hecoinfo)
+heco_txns = DataParser(
+        DataParser.TYPE_EXPLORER,
+        "HecoInfo (HECO Transactions)",
+        ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
+         'Value_IN(HT)', 'Value_OUT(HT)', None, 'TxnFee(HT)', 'TxnFee(USD)',
+         'Historical $Price/HT', 'Status', 'ErrCode'],
+        worksheet_name="HecoInfo",
+        row_handler=parse_hecoinfo)
 
-heco2 = DataParser(DataParser.TYPE_EXPLORER,
+DataParser(DataParser.TYPE_EXPLORER,
            "HecoInfo (HECO Transactions)",
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
             'Value_IN(HT)', 'Value_OUT(HT)', None, 'TxnFee(HT)', 'TxnFee(USD)',
@@ -71,7 +79,7 @@ heco2 = DataParser(DataParser.TYPE_EXPLORER,
            worksheet_name="HecoInfo",
            row_handler=parse_hecoinfo)
 
-heco_internal = DataParser(DataParser.TYPE_EXPLORER,
+DataParser(DataParser.TYPE_EXPLORER,
            "HecoInfo (HECO Internal Transactions)",
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'ParentTxFrom', 'ParentTxTo',
             'ParentTxETH_Value', 'From', 'TxTo', 'ContractAddress', 'Value_IN(HT)', 'Value_OUT(HT)',
@@ -79,23 +87,13 @@ heco_internal = DataParser(DataParser.TYPE_EXPLORER,
            worksheet_name="HecoInfo",
            row_handler=parse_hecoinfo_internal)
 
-heco_internal = DataParser(DataParser.TYPE_EXPLORER,
+DataParser(DataParser.TYPE_EXPLORER,
            "HecoInfo (HECO Internal Transactions)",
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'ParentTxFrom', 'ParentTxTo',
             'ParentTxETH_Value', 'From', 'TxTo', 'ContractAddress', 'Value_IN(HT)', 'Value_OUT(HT)',
             None, 'Historical $Price/HT', 'Status', 'ErrCode', 'Type', 'PrivateNote'],
            worksheet_name="HecoInfo",
            row_handler=parse_hecoinfo_internal)
-
-def find_same_tx(data_file, tx_hash):
-    for data_row in data_file.data_rows:
-        if data_row.t_record and data_row.row_dict['Txhash'] == tx_hash:
-            fee_quantity = data_row.t_record.fee_quantity
-            fee_asset = data_row.t_record.fee_asset
-            data_row.t_record = None
-            return fee_quantity, fee_asset
-
-    return None, ''
 
 # Same header as Etherscan
 #DataParser(DataParser.TYPE_EXPLORER,